backend/requirements.txt
asgiref==3.5.2
Django==4.1.6
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
inflection==0.5.1
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
psycopg2-binary==2.9.5
django-cors-headers==3.13.0
